Add keycloak integration to virtual lab

+++ virtual_labs/usecases/labs/delete_virtual_lab.py
@@ -8,7 +8,7 @@ from sqlalchemy.orm import Session
 from virtual_labs.core.exceptions.api_error import VliError, VliErrorCode
 from virtual_labs.infrastructure.db import models
 from virtual_labs.repositories import labs as repository
-from virtual_labs.repositories.group_repo import GroupMutationRepository
+from virtual_labs.usecases.labs.lab_authorization import is_user_admin_of_lab
 
 
 class ProjectWithIds(BaseModel):
@@ -21,31 +21,24 @@ class ProjectWithIds(BaseModel):
         from_attributes = True
 
 
-def delete_virtual_lab(db: Session, lab_id: UUID4) -> models.VirtualLab:
+def delete_virtual_lab(db: Session, lab_id: UUID4, user_id: UUID4) -> models.VirtualLab:
     try:
         lab = repository.get_virtual_lab(db, lab_id)
-        group_repo = GroupMutationRepository()
-
-        # Delete groups for lab
-        group_repo.delete_group(group_id=str(lab.admin_group_id))
-        group_repo.delete_group(group_id=str(lab.member_group_id))
-
-        # Delete groups for un-deleted children projects also
-        for db_project in lab.projects:
-            project = ProjectWithIds.model_validate(db_project)
-            if not project.deleted:
-                group_repo.delete_group(group_id=project.admin_group_id)
-                group_repo.delete_group(group_id=project.member_group_id)
-
-        # Now mark db row as deleted
-        repository.delete_virtual_lab(db, lab_id)
-        return lab
+        if not is_user_admin_of_lab(user_id, lab):
+            raise VliError(
+                message=f"Only admins of virtual lab {lab.name} are authorized to delete virtual lab.",
+                error_code=VliErrorCode.NOT_ALLOWED_OP,
+                http_status_code=HTTPStatus.FORBIDDEN,
+            )
+        return repository.delete_virtual_lab(db, lab_id)
     except SQLAlchemyError:
         raise VliError(
             message="Virtual lab not found",
             error_code=VliErrorCode.ENTITY_NOT_FOUND,
             http_status_code=HTTPStatus.NOT_FOUND,
         )
+    except VliError as error:
+        raise error
     except Exception as error:
         logger.warning(f"Deleting virtual lab groups failed  failed: {error}")
         raise VliError(
